update auth and add edit

+++ app/Http/Controllers/ArticleController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
+use App\Models\Category;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -36,10 +38,7 @@ class ArticleController extends Controller
 
     public function add()
     {
-        $data = [
-            ['id' => 1, 'name' => 'News'],
-            ['id' => 2, 'name' => "Tech"],
-        ];
+        $data = Category::all();
 
         return view('articles.add', [
             'categories' => $data,
@@ -63,6 +62,7 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = auth()->user()->id;
         $article->save();
 
         return redirect("/articles");
@@ -71,8 +71,45 @@ class ArticleController extends Controller
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
+        if(Gate::allows('article-delete', $article)) {
+            $article->delete();
+        } else {
+            return redirect("/articles");
 
-        return redirect("/articles")->with('info', "An Article \"$article->title\" is Deleted.");
+        }
+    }
+
+    public function edit($id)
+    {
+        $article = Article::find($id);
+        $categories = Category::all();
+
+        return view('articles.edit', [
+            'article' => $article,
+            'categories' => $categories
+        ]);
+    }
+
+    public function update($id)
+    {
+        $validator = validator(request()->all(), [
+            'title' => 'required',
+            'body' => 'required',
+            'category_id' => 'required',
+        ]);
+
+        if ($validator->fails()) {
+            return back()->withErrors($validator);
+        }
+
+        $article = Article::find($id);
+
+        $article->title = request()->title;
+        $article->body = request()->body;
+        $article->category_id = request()->category_id;
+        $article->user_id = auth()->user()->id;
+        $article->save();
+
+        return redirect("/articles");
     }
 }
